Rename file_watcher -> path_watcher and reexport it
