+++ lib/streamlit/bootstrap.py
@@ -33,8 +33,7 @@ from streamlit.config import CONFIG_FILENAMES
 from streamlit.logger import get_logger
 from streamlit.secrets import SECRETS_FILE_LOC
 from streamlit.server.server import Server, server_address_is_unix_socket
-from streamlit.watcher.file_watcher import watch_file
-from streamlit.watcher.file_watcher import report_watchdog_availability
+from streamlit.watcher import report_watchdog_availability, watch_file
 
 LOGGER = get_logger(__name__)
